BUILDSYS-321 fixed build on mac

+++ cpp-lib/include/SNACCDeprecated.h
@@ -32,7 +32,7 @@ public:
 	 * @param szObjectName - the name of the object that is about to get created
 	 * @param callStack - the call stack that shows where the object has been created
 	 */
-	virtual void DeprecatedASN1Object(const __int64 i64DeprecatedSince, const char* szModuleName, const char* szObjectName, const std::list<std::string>& callStack) = 0;
+	virtual void DeprecatedASN1Object(const long long i64DeprecatedSince, const char* szModuleName, const char* szObjectName, const std::list<std::string>& callStack) = 0;
 
 	/**
 	 * A deprecated method has been called
@@ -44,7 +44,7 @@ public:
 	 * @param callStack - the call stack that shows where the object has been created
 	 * @param pContext - The invoke Context that is associated with the invoke
 	 */
-	virtual void DeprecatedASN1Method(const __int64 i64DeprecatedSince, const char* szModuleName, const char* szMethodName, const SNACCDeprecatedNotifyCallDirection direction, const std::list<std::string>& callStack, const SnaccInvokeContext* pContext = NULL) = 0;
+	virtual void DeprecatedASN1Method(const long long i64DeprecatedSince, const char* szModuleName, const char* szMethodName, const SNACCDeprecatedNotifyCallDirection direction, const std::list<std::string>& callStack, const SnaccInvokeContext* pContext = NULL) = 0;
 };
 
 class SNACCDeprecated
@@ -64,7 +64,7 @@ public:
 	 * @param szModuleName - the module in which the object is located
 	 * @param szObjectName - the name of the object that is about to get created
 	 */
-	static void DeprecatedASN1Object(const __int64 i64DeprecatedSince, const char* szModuleName, const char* szObjectName);
+	static void DeprecatedASN1Object(const long long i64DeprecatedSince, const char* szModuleName, const char* szObjectName);
 
 	/**
 	 * This method is called in case a method is called which is flagged deprecated
@@ -75,7 +75,7 @@ public:
 	 * @param direction - whether the call was inbound or outbound
 	 * @param pContext - the invokeContext that shows more details about the invoke
 	 */
-	static void DeprecatedASN1Method(const __int64 i64DeprecatedSince, const char* szModuleName, const char* szMethodName, const SNACCDeprecatedNotifyCallDirection direction, const SnaccInvokeContext* pContext = NULL);
+	static void DeprecatedASN1Method(const long long i64DeprecatedSince, const char* szModuleName, const char* szMethodName, const SNACCDeprecatedNotifyCallDirection direction, const SnaccInvokeContext* pContext = NULL);
 
 private:
 	/**
